+++ ESSArch_Core/storage/filters.py
@@ -39,6 +39,17 @@ class StorageMediumFilter(filters.FilterSet):
         field_name='storage_target__storage_method_target_relations__storage_method__type',
         choices=storage_type_CHOICES
     )
+    deactivatable = filters.BooleanFilter(label='deactivatable', method='filter_deactivatable')
+    include_inactive_ips = filters.BooleanFilter(method='filter_include_inactive_ips')
+
+    def filter_include_inactive_ips(self, queryset, *args):
+        # this filter is only used together with deactivatable
+        return queryset
+
+    def filter_deactivatable(self, queryset, name, value):
+        include_inactive_ips = self.request.query_params.get('include_inactive_ips', False)
+        include_inactive_ips = include_inactive_ips in (True, 'True', 'true', '1')
+        return queryset.deactivatable(include_inactive_ips=include_inactive_ips)
 
     class Meta:
         model = StorageMedium